feat: (#499) SchoolResponse
