--- dms-core/src/main/kotlin/team/aliens/dms/domain/school/dto/NeisSchoolResponse.kt
@@ -1,6 +1,0 @@
-package team.aliens.dms.domain.school.dto
-
-data class NeisSchoolResponse(
-    val sdSchoolCode: String,
-    val regionCode: String
-)
