--- dms-core/src/main/kotlin/team/aliens/dms/domain/school/dto/QueryAvailableFeaturesResponse.kt
@@ -1,9 +1,0 @@
-package team.aliens.dms.domain.school.dto
-
-data class QueryAvailableFeaturesResponse(
-    val mealService: Boolean,
-    val noticeService: Boolean,
-    val pointService: Boolean,
-    val studyRoomService: Boolean,
-    val remainService: Boolean
-)
