--- dms-presentation/src/main/kotlin/team/aliens/dms/domain/school/dto/response/ReissueSchoolCodeResponse.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.school.dto.response
-
-data class ReissueSchoolCodeResponse(
-    val code: String
-)
