--- dms-presentation/src/main/kotlin/team/aliens/dms/domain/school/dto/response/SchoolIdResponse.kt
@@ -1,7 +1,0 @@
-package team.aliens.dms.domain.school.dto.response
-
-import java.util.UUID
-
-data class SchoolIdResponse(
-    val schoolId: UUID
-)
