--- dms-presentation/src/main/kotlin/team/aliens/dms/domain/school/dto/response/SchoolQuestionResponse.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.school.dto.response
-
-data class SchoolQuestionResponse(
-    val question: String
-)
